gameLauncher 로 클래스명 변경 , move 명령 추가
